Compiler happy. Developer unhappy

+++ src/Storages/StorageGenerateRandom.cpp
@@ -706,7 +706,7 @@ Pipe StorageGenerateRandom::read(
     if (query_limit && num_streams * max_block_size > query_limit)
     {
         /// We want to avoid spawning more streams than necessary
-        num_streams = std::min(num_streams, ((query_limit + max_block_size - 1) / max_block_size));
+        num_streams = std::min(num_streams, static_cast<size_t>(((query_limit + max_block_size - 1) / max_block_size)));
     }
     Pipes pipes;
     pipes.reserve(num_streams);
